refactors making API calls
