+++ mobile-app/mobile_app/lib/config/globals.dart
@@ -1,0 +1,5 @@
+library globals;
+
+import 'package:http/http.dart' as http;
+
+http.Client httpClient = http.Client();
